refactor: clean up first commit
