+++ src/pages/auth/profile/styled/avatarContainer.js
