+++ src/pages/auth/profile/styled/bottomBar.js
