+++ src/pages/public/reset-password-confirmation/icons/mailSendIcon.js
