+++ src/pages/public/terms-condition/styles.css
