+++ src/routes.js
@@ -8,8 +8,8 @@ import PropTypes from 'prop-types';
 import history from './history';
 import AuthRoute from './components/experiences/auth-route';
 import PublicRoute from './components/experiences/public-route';
-import PublicRoutes from './components/pages/public';
-import AuthRoutes from './components/pages/auth';
+import PublicRoutes from './pages/public';
+import AuthRoutes from './pages/auth';
 
 export default function Routes({ isAuth }) {
     return (
